requirements.txt
aws-cdk-lib==2.52.0
cdk-nag==2.21.73
constructs>=10.0.0,<11.0.0
